nps_apis/requirements.txt
absl-py==0.12.0
anyio==3.3.4
asgiref==3.3.4
astunparse==1.6.3
blis==0.7.4
cached-property==1.5.2
cachetools==4.2.2
catalogue==2.0.4
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cloudpickle==2.0.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cymem==2.0.5
dask==2021.9.1
dill==0.3.3
Django==3.2.2
django-cors-headers==3.10.0
django-reversion==1.8.7
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
drf-yasg==1.17.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0-py3-none-any.whl
Faker==8.1.3
fastapi==0.70.0
filelock==3.0.12
flatbuffers==1.12
fsspec==2021.10.1
future==0.18.2
gast==0.3.3
google-auth==1.30.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.32.0
h5py==2.10.0
idna==2.10
importlib-metadata==4.0.1
inflection==0.5.1
iohandler==1.0.6
itypes==1.2.0
Jinja2==2.11.3
joblib==1.0.1
Keras==2.3.1
Keras-Applications==1.0.8
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
locket==0.2.1
lxml==4.7.1
Markdown==3.3.4
MarkupSafe==1.1.1
murmurhash==1.0.5
mysqlclient==2.1.0
networkx==2.6.3
nlp==0.4.0
nltk==3.6.2
numpy==1.19.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.4
pandas==1.3.5
partd==1.2.0
pathlib==1.0.1
pathy==0.5.2
pke==1.8.1
preshed==3.0.5
protobuf==3.17.0
psycopg2==2.9.1
pyarrow==4.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.7.4
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==6.0
rake-nltk==1.0.6
regex==2021.4.4
renderer==1.0.3
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
scikit-learn==1.0
scipy==1.4.1
six==1.15.0
sklearn==0.0
smart-open==3.0.0
sniffio==1.2.0
spacy==3.0.6
spacy-legacy==3.0.5
SQLAlchemy==1.4.25
sqlparse==0.4.1
srsly==2.4.1
starlette==0.16.0
structlog==21.1.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.0
tensorflow-estimator==2.4.0
termcolor==1.1.0
text-unidecode==1.3
textblob==0.15.3
thinc==8.0.3
threadpoolctl==3.0.0
toolz==0.11.1
tqdm==4.61.0
typer==0.3.2
typing==3.7.4.3
typing-extensions==3.7.4.3
Unidecode==1.3.2
uritemplate==3.0.1
urllib3==1.26.4
wasabi==0.8.2
Werkzeug==2.0.1
wrapt==1.12.1
xxhash==2.0.2
zipp==3.4.1
